climbing-guidebook/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
redis==5.0.1
pillow==10.1.0
python-multipart==0.0.6
httpx==0.25.1